Poject_B. Schedule new broker. Tunning competition detect algo. Larger type for broker.

+++ Code/Project_B/CodeServerSide/DataProvider/BetProvider.cs
@@ -41,7 +41,7 @@ namespace Project_B.CodeServerSide.DataProvider {
                 var betWithAdvancedDb = Bet.DataSource
                     .Join(JoinType.Left, BetAdvanced.Fields.BetID, Bet.Fields.ID, RetrieveMode.Retrieve)
                     .WhereEquals(Bet.Fields.CompetitionitemID, competitionItemID)
-                    .WhereEquals(Bet.Fields.BrokerID, (short) brokerType)
+                    .WhereEquals(Bet.Fields.BrokerID, (int) brokerType)
                     .Sort(Bet.Fields.ID, SortDirection.Desc)
                     .First();
                 var betAdvancedDb = betWithAdvancedDb != null ? betWithAdvancedDb.GetJoinedEntity<BetAdvanced>() : null;
@@ -86,7 +86,7 @@ namespace Project_B.CodeServerSide.DataProvider {
                 var state = _betTypeByHour[(short)dateUtc.Hour];
                 var systemState = SystemStateBet.DataSource
                     .WhereEquals(SystemStateBet.Fields.Dateutc, dateUtc.Date)
-                    .WhereEquals(SystemStateBet.Fields.Brokerid, (short) brokerType)
+                    .WhereEquals(SystemStateBet.Fields.Brokerid, (int) brokerType)
                     .First();
                 if (systemState == null) {
                     systemState = new SystemStateBet {
@@ -105,7 +105,7 @@ namespace Project_B.CodeServerSide.DataProvider {
                 var state = _betTypeByHour[(short)dateUtc.Hour];
                 var systemState = SystemStateBet.DataSource
                     .WhereEquals(SystemStateBet.Fields.Dateutc, dateUtc.Date)
-                    .WhereEquals(SystemStateBet.Fields.Brokerid, (short) brokerType)
+                    .WhereEquals(SystemStateBet.Fields.Brokerid, (int) brokerType)
                     .First();
                 if (systemState == null) {
                     systemState = new SystemStateBet {
