+++ Code/Project_B/CodeServerSide/Entity/SystemStateResult.cs
@@ -34,7 +34,7 @@ namespace Project_B.CodeServerSide.Entity {
             /// <summary>
             /// 
             /// </summary>
-            [DBField(DbType.Int16)]
+            [DBField(DbType.Int32)]
             BrokerID,
 
             /// <summary>
@@ -90,7 +90,7 @@ namespace Project_B.CodeServerSide.Entity {
         /// 
         /// </summary>
         public BrokerType BrokerID {
-            get { return (BrokerType)(short)this[Fields.BrokerID]; }
+            get { return (BrokerType)(int)this[Fields.BrokerID]; }
             set { ForceSetData(Fields.BrokerID, value); }
         }
